refactor: [#1182] move structs to new mods in whitelist
